+++ checkpoint/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Enable `ArrayMetadata` persistence globally.
+
 ## [0.11.7] - 2025-03-06
 
 ### Added